Feat(node): initial module structure
