+++ src/main/scala/hydrozoa/Main.scala
@@ -1,12 +1,14 @@
 package hydrozoa
 
 import com.typesafe.scalalogging.Logger
-import hydrozoa.head.l1.AppCtx.yaciDevKit
-import hydrozoa.head.l1.txbuilder.{BloxBeanTxBuilder, TxBuilder}
-import hydrozoa.head.l1.{AppCtx, Cardano, YaciDevKitCardano}
-import hydrozoa.head.network.{HydrozoaNetwork, MockHydrozoaNetwork}
-import hydrozoa.head.wallet.{MockWallet, Wallet}
-import hydrozoa.head.{Node, NodeApi, genNodeKey}
+import hydrozoa.infra.genNodeKey
+import hydrozoa.l1.multisig.tx.initialization.{BloxBeanInitTxBuilder, InitTxBuilder}
+import hydrozoa.l1.wallet.{MockWallet, Wallet}
+import hydrozoa.l1.{Cardano, YaciDevKitCardano}
+import hydrozoa.l2.consensus.network.{HydrozoaNetwork, MockHydrozoaNetwork}
+import AppCtx.yaciDevKit
+import hydrozoa.node.api.NodeApi
+import hydrozoa.node.server.Node
 
 object Cli:
 
@@ -19,7 +21,7 @@ object Cli:
         val log: Logger = Logger("Hydrozoa")
         val wallet: Wallet = MockWallet(ctx)
         val cardano: Cardano = YaciDevKitCardano(ctx)
-        val txBuilder: TxBuilder = BloxBeanTxBuilder(ctx)
+        val txBuilder: InitTxBuilder = BloxBeanInitTxBuilder(ctx)
         val network: HydrozoaNetwork = MockHydrozoaNetwork(txBuilder, cardano, ownKeys._2)
         val node = Node(ownKeys, network, cardano, wallet, txBuilder, log)
