--- src/main/scala/hydrozoa/head/HeadParams.scala
@@ -1,5 +1,0 @@
-package hydrozoa.head
-
-import com.bloxbean.cardano.client.common.model.Network
-
-case class HeadParams(networkId: Network, participants: Set[ParticipantVerificationKey])
