--- src/main/scala/hydrozoa/head/l1/AppCtx.scala
@@ -1,35 +1,0 @@
-package hydrozoa.head.l1
-
-import com.bloxbean.cardano.client.account.Account
-import com.bloxbean.cardano.client.backend.api.BackendService
-import com.bloxbean.cardano.client.backend.blockfrost.service.BFBackendService
-import com.bloxbean.cardano.client.common.model.Network
-import scalus.builtin.ByteString
-import scalus.ledger.api.v1.PubKeyHash
-
-// TODO: AppCtx in fact just holds backendService for now
-
-case class AppCtx(
-    network: Network,
-    account: Account,
-    backendService: BackendService
-) {
-    lazy val pubKeyHash: PubKeyHash = PubKeyHash(
-      ByteString.fromArray(account.hdKeyPair().getPublicKey.getKeyHash)
-    )
-}
-
-object AppCtx {
-
-    def yaciDevKit(): AppCtx = {
-        val url = "http://localhost:8080/api/v1/"
-        val network = new Network(0, 42)
-        val mnemonic =
-            "test test test test test test test test test test test test test test test test test test test test test test test sauce"
-        AppCtx(
-          network,
-          new Account(network, mnemonic),
-          new BFBackendService(url, "")
-        )
-    }
-}
