--- src/main/scala/hydrozoa/head/l1/Cardano.scala
@@ -1,10 +1,0 @@
-package hydrozoa.head.l1
-
-import hydrozoa.head.{Network, L1Tx, TxId}
-
-trait Cardano {
-    def submit(tx: L1Tx): Either[SubmissionError, TxId]
-    def network(): Network
-}
-
-type SubmissionError = String
