--- src/main/scala/hydrozoa/head/l1/txbuilder/TxBuilder.scala
@@ -1,16 +1,0 @@
-package hydrozoa.head.l1.txbuilder
-
-import hydrozoa.head.*
-
-trait TxBuilder {
-    def mkInitDraft(recipe: InitTxRecipe): Either[String, L1Tx]
-}
-
-case class InitTxRecipe(
-    headAddressBech32: String,
-    txId: TxId,
-    txIx: TxIx,
-    amount: Long,
-    headNativeScript: NativeScript,
-    beaconTokenName: String
-)
