--- src/main/scala/hydrozoa/head/wallet/Wallet.scala
@@ -1,7 +1,0 @@
-package hydrozoa.head.wallet
-
-import hydrozoa.head.{TxKeyWitness, L1Tx}
-
-trait Wallet {
-    def sign(tx: L1Tx): TxKeyWitness
-}
