+++ src/main/scala/hydrozoa/l1/YaciDevKitCardano.scala
@@ -1,0 +1,17 @@
+package hydrozoa.l1
+
+import hydrozoa.{AppCtx, L1Tx, Network, TxId}
+
+class YaciDevKitCardano(ctx: AppCtx) extends Cardano {
+
+    private val backendService = ctx.backendService
+
+    override def submit(tx: L1Tx): Either[SubmissionError, TxId] = {
+        val result = backendService.getTransactionService.submitTransaction(tx.bytes)
+        Either.cond(result.isSuccessful, TxId(result.getValue), result.getResponse)
+    }
+
+    override def network(): Network =
+        val nw = ctx.network
+        Network(nw.getNetworkId, nw.getProtocolMagic)
+}
