+++ src/main/scala/hydrozoa/l1/wallet/Wallet.scala
@@ -1,0 +1,7 @@
+package hydrozoa.l1.wallet
+
+import hydrozoa.{L1Tx, TxKeyWitness}
+
+trait Wallet {
+    def sign(tx: L1Tx): TxKeyWitness
+}
