+++ src/main/scala/hydrozoa/l2/consensus/HeadParams.scala
@@ -1,0 +1,8 @@
+package hydrozoa.l2.consensus
+
+import hydrozoa.{Network, ParticipantVerificationKey}
+
+case class HeadParams(
+    networkId: Network,
+    participants: Set[ParticipantVerificationKey]
+)
